........
Feature #820
Comments
........
Feature #820
Comments
........
Feature #820
Comments
........
Feature #820
Comments
........
Feature #820
Comments
........
Feature #820
Comments
........
Feature #820
Comments
........
Feature #820
Comments
........
Feature #816: Added missing function getSize
........
Feature #816: Fixed warnings
........
Feature #816: Fixed warnings
........
Feature #817: Updated comments
........
Feature #816: Compile commandline optimisation
........
Feature #820
Added info about missing feature
Removed "_tag*" names
........
Feature #820
Removed "_tag*" names
........
Feature #820
Comments updated
........

........

........
Feature #817: update doxygen comments



........
Feature #817: Added logo
........
Feature #817: Doygen syntax (no typo, but inconsistent names
........
Feature #817: Comments in examples
........
Feature #817: Adde Qt example
........
Feature #817: Comments
........
Feature #817: 
Hide some defines from docs
Do not show hide commands in docs

........
Bug #822: remove VS 2010 project from dist
........
Feature #817: update doxygen comments 
Feature #816: change function name for getErrorString



........

........
Feature #817: 
/endcond etc. did not work

........
Refactoring
........
Feature #816: Declare const pointers were appropriate
........
Feature #816: 
Declare const pointers were appropriate
Fixed PTT values (forgot BE2LE)
........
Feature #816: 
DVDPTT program chain number was duplicate
........
Bug #829:
Added VMG_PTT_SRPT structure
Added Physical/Virtual view check box
........
Feature #817: updated info
........
Feature #828: Allow switching physical/virtual view
Bug #829: add support for PTT
........
Test code
........
Test code
........
Bug #831: Last chapter/program had no cells
........
Bug #831: Last chapter/program had no cells
........

........

+++ lib/dvdutils.cpp
@@ -30,7 +30,7 @@
 
 DLL_PUBLIC uint64_t dvdSector2bytes(uint32_t dwSectorAddress)
 {
-    return ((uint64_t)DVD_SECTOR_SIZE * dwSectorAddress);
+    return ((uint64_t)DVD_SECTOR_SIZE * (uint64_t)dwSectorAddress);
 }
 
 DLL_PUBLIC const char * dvdGetExtension(DVDFILETYPE eFileType)
